OIDC Dynamic Client Registration endpoint (#4)
The registration endpoint is only made available for businesses with a custom domain. This is ideal in enterprise scenarios where a company has to register many clients, and can do so programmatically, or even letting the client (a third party app) self-register to their private OIDC IdP.

The endpoint is protected by a Bearer token, as per specification, which is unique per tenant (our business entity). I suppose later on, this can become a lot more elaborate, with API keys, restricted scopes etc.

The reason why the self-registration endpoint is not available on the public OIDC IdP is simply: spam. We will likely have to review apps signing up, making sure the scopes requested make sense and are not abused.