Relatively production-like infrastructure based on K8s on Hetzner
Defines a golden image with Ansible + Packer.
Creates a private Network, with a NAT gateway.
Auto-configures instances in the private network to use the nat gateway and public DNS.
Defines a Kubernetes cluster of 3 master nodes (1 in Finland), 1 monitoring node, 2 ingress nodes.
Uses Celium + WireGuard for encrypted pod networking
The ingresses are hooked up to cert-manager with Let's Encrypt for auto cert management.
Installs CloudNativePG, launches a 3 node cluster (co-located on the masters).
The 3 nodes have encrypted ZFS storage, with Tang/Clevis for decryption at boot.
One node is in finland, so sync replication is "discouraged".