OIDC vs OAuth 2.0 (#32)
The OIDC specification requires the openid scope to be passed in the Authorization request. However, when that's missing, we don't necessarily have to return an error.

This PR adjusts the authorization logic to allow noo id to work as an OAuth 2.0 Server as well. This is useful when third party apps require API access to customer resources without necessarily needing to authenticate the user.

When the openid scope is not passed, the request is "downgraded". An id_token will not be granted, and the access token resulting from the flow will not have access to the userinfo endpoint.