Rewrite the id app using the react-router framework.
This large rewrite was due to all the instabilities and weirdnesses of
Next.js.

Some code has been moved into the @noo/lib kit, and the oidc-server
library has been updated to allow to be compatible with more than just
Next.js.

A lot of code has been rewritten, and there's plenty to refactor.

Bun has been replaced by pnpm as package manager, and by node.js as
runner. I love Bun, but in large monorepos it is just not
production-ready yet (and it also causes some troubles with
react-router).

This should be functionally equivalent to the previous app.